requirements.txt
numpy
opencv-python
Pillow
torch>=1.7
torchvision>=0.8.0
huggingface-hub
basicsr>=1.4.2
facexlib>=0.2.5
gfpgan>=1.3.5
ffmpeg-python
tqdm
